Fantasy points and Ranks

Fantasy points and Ranks

fp_fantasy_points(pos = "overall", season, start_week = 1,
  end_week = 17, scoring = c("half", "ppr", "std"))

Arguments

pos
Charcater. Specific position you want to return. Default will return all all offensive postions. Available options include

"overall"

"QB"

"RB"

"WR"

"TE"

"K"

"DST"

"IDP"

"DL"

"LB"

"DB"

season
Numeric. The NFL season. If missing it will return the current year's season. Supported season only go back to 2015

start_week
Numeric. The starting week. Default is 1

end_week
Numeric. The ending week. Default is 17

scoring
Charcater. Fantasy scoring format. Default is "half"

Value

a tibble

Note

If you query a single week all players that were on a bye that week are not returned

Examples

# overall fantasy points for the entire season
fp_fantasy_points(season = 2018)

#> Registered S3 method overwritten by 'rvest':
#>   method            from
#>   read_xml.response xml2

#> # A tibble: 669 x 11
#>     rank player team  position season start_week end_week scoring points games
#>    <dbl> <chr>  <chr> <chr>     <dbl>      <dbl>    <dbl> <chr>    <dbl> <dbl>
#>  1     1 Patri~ KC    QB         2018          1       17 half      417     16
#>  2     2 Matt ~ ATL   QB         2018          1       17 half      354.    16
#>  3     3 Todd ~ LAR   RB         2018          1       17 half      343.    14
#>  4     4 Ben R~ PIT   QB         2018          1       17 half      342.    16
#>  5     5 Saquo~ NYG   RB         2018          1       17 half      340.    16
#>  6     6 Chris~ CAR   RB         2018          1       17 half      332     16
#>  7     7 Desha~ HOU   QB         2018          1       17 half      332.    16
#>  8     8 Andre~ IND   QB         2018          1       17 half      328.    16
#>  9     9 Alvin~ NO    RB         2018          1       17 half      314.    15
#> 10    10 Aaron~ GB    QB         2018          1       17 half      312.    16
#> # ... with 659 more rows, and 1 more variable: avg <dbl>

# TE's in the 2017 season using standard scoring
fp_fantasy_points(pos = "TE", season = 2017, scoring = "std")

#> Error: Can't find column `position` in `.data`.

fp_fantasy_points(pos = "IDP", season = 2015)

#> # A tibble: 915 x 11
#>     rank player team  position season start_week end_week scoring points games
#>    <dbl> <chr>  <chr> <chr>     <dbl>      <dbl>    <dbl> <chr>    <dbl> <dbl>
#>  1     1 Resha~ MIA   S          2015          1       17 half      279.    16
#>  2     2 Lavon~ TB    LB         2015          1       17 half      276.    16
#>  3     3 D'Qwe~ FA    LB         2015          1       17 half      266.    16
#>  4     4 J.J. ~ HOU   DE         2015          1       17 half      266.    16
#>  5     5 Telvi~ JAC   LB         2015          1       17 half      252.    14
#>  6     6 Paul ~ FA    LB         2015          1       17 half      249.    14
#>  7     7 Navor~ FA    LB         2015          1       17 half      246.    16
#>  8     8 Malco~ SF    LB         2015          1       17 half      245.    16
#>  9     9 Khali~ CHI   LB         2015          1       17 half      236.    15
#> 10    10 Luke ~ CAR   LB         2015          1       17 half      236.    13
#> # ... with 905 more rows, and 1 more variable: avg <dbl>