Detailed Analysis of NFL snap counts

Detailed Analysis of NFL snap counts

fp_snap_analysis(pos = "offense", season, start_week = 1,
  end_week = 17, scoring = c("half", "ppr", "std"))

Arguments

pos
Charcater. Specific position you want to return. Default will return all all offensive postions. Available options include

"offense"

"QB"

"RB"

"WR"

"TE"

"Defense"

"DL"

"LB"

"DB"

season
Numeric. The NFL season. If missing it will return the current year's season. Supported season only go back to 2016

start_week
Numeric. The starting week. Default is 1

end_week
Numeric. The ending week. Default is 17

scoring
Charcater. Fantasy scoring format. Default is "half"

Value

a tibble

Note

If you query a single week all players that were on a bye that week are not returned

Examples

# all offensive positions for weeks 5-9 of the 2018 season
fp_snap_analysis(season = 2018, start_week = 5, end_week = 9)

#> # A tibble: 423 x 17
#>    player pos   team  season start_week end_week scoring games snaps snaps_gm
#>    <chr>  <chr> <chr>  <dbl>      <dbl>    <dbl> <chr>   <dbl> <dbl>    <dbl>
#>  1 Aaron~ QB    GB      2018          5        9 half        4   278       70
#>  2 Adria~ RB    WAS     2018          5        9 half        5   157       31
#>  3 Alex ~ QB    WAS     2018          5        9 half        5   328       66
#>  4 Ben R~ QB    PIT     2018          5        9 half        4   283       71
#>  5 Benja~ TE    NO      2018          5        9 half        4   119       30
#>  6 Brian~ QB    NE      2018          5        9 half        1     2        2
#>  7 Chase~ QB    CHI     2018          5        9 half        2     4        2
#>  8 Danny~ WR    MIA     2018          5        9 half        5   295       59
#>  9 DeSea~ WR    TB      2018          5        9 half        4   155       39
#> 10 Drew ~ QB    NO      2018          5        9 half        4   258       65
#> # ... with 413 more rows, and 7 more variables: snap_percent <dbl>,
#> #   rush_percent <dbl>, tgt_percent <dbl>, touch_percent <dbl>,
#> #   util_percent <dbl>, fantasy_pts <dbl>, pts_100_snaps <dbl>

# TE's in the 2017 season using standard scoring
fp_snap_analysis(pos = "TE", season = 2017, scoring = "std")

#> # A tibble: 80 x 16
#>    player team  season start_week end_week scoring games snaps snaps_gm
#>    <chr>  <chr>  <dbl>      <dbl>    <dbl> <chr>   <dbl> <dbl>    <dbl>
#>  1 Benja~ BAL     2017          1       17 std        16   699       44
#>  2 Delan~ TEN     2017          1       17 std        16   744       47
#>  3 Greg ~ CAR     2017          1       17 std         7   367       52
#>  4 Jared~ OAK     2017          1       17 std        16   796       50
#>  5 Jason~ DAL     2017          1       17 std        16  1050       66
#>  6 Marce~ JAC     2017          1       17 std        16   895       56
#>  7 Verno~ WAS     2017          1       17 std        16   803       50
#>  8 Ed Di~ CAR     2017          1       17 std        16   875       55
#>  9 Logan~ SF      2017          1       17 std        14   147       11
#> 10 Kyle ~ MIN     2017          1       17 std        16   924       58
#> # ... with 70 more rows, and 7 more variables: snap_percent <dbl>,
#> #   rush_percent <dbl>, tgt_percent <dbl>, touch_percent <dbl>,
#> #   util_percent <dbl>, fantasy_pts <dbl>, pts_100_snaps <dbl>

# all defensive positions for the current NFL season in week 1
fp_snap_analysis(pos = "defense", start_week = 1, end_week = start_week)

#> Error in fp_build_query_list(season = season, scoring = scoring, start_week = start_week,     end_week = end_week, snaps = 0): object 'start_week' not found