NFL Snap Counts by Week

NFL Snap Counts by Week

fp_snap_counts(pos = "offense", season, percentage = FALSE)

Arguments

pos
Charcater. Specific position you want to return. Default will return all all offensive postions. Available options include

"offense"

"QB"

"RB"

"WR"

"TE"

"Defense"

"DL"

"LB"

"DB"

season
Numeric. The NFL season. If missing it will return the current year's season. Supported season only go back to 2016

percentage
Logical. Return counts as percentages? Default is FALSE

Value

a tibble

Note

Due to how fantasypros returns the data, all player's teams are listed under their current team regardless of what season you query.

Bye weeks are represented as NA

Examples

# all offensive positions for the 2018 season
fp_snap_counts(season = 2018)

#> # A tibble: 505 x 23
#>    player pos   team  season    w1    w2    w3    w4    w5    w6    w7    w8
#>    <chr>  <chr> <chr>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Aaron~ QB    GB      2018    46    77    69    76    81    71    NA    52
#>  2 Adria~ RB    WAS     2018    42    25    32    NA    13    37    34    41
#>  3 Alex ~ QB    WAS     2018    79    74    61    NA    61    70    60    68
#>  4 Ben R~ QB    PIT     2018    84    82    66    62    60    73    NA    70
#>  5 Benja~ TE    NE      2018    51    54    45    37    36     0    36    14
#>  6 Brian~ QB    NE      2018     0     0     0     4     0     0     0     2
#>  7 Chad ~ QB    KC      2018     0     0     0     0     0     0     0     0
#>  8 Chase~ QB    CHI     2018     0     0     0     3    NA     0     1     0
#>  9 Danny~ WR    DET     2018    45    44    31    42    53    75    55    65
#> 10 Delan~ TE    TEN     2018    39     0     0     0     0     0     0    NA
#> # ... with 495 more rows, and 11 more variables: w9 <dbl>, w10 <dbl>,
#> #   w11 <dbl>, w12 <dbl>, w13 <dbl>, w14 <dbl>, w15 <dbl>, w16 <dbl>,
#> #   w17 <dbl>, ttl <dbl>, avg <dbl>

# Percentage of total game snaps for TE's in the 2018 season
fp_snap_counts(pos = "TE", season = 2018, percentage = TRUE)

#> # A tibble: 110 x 22
#>    player team  season    w1    w2    w3    w4    w5    w6    w7    w8    w9
#>    <chr>  <chr>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Benja~ NE      2018    80    82    57    53    55     0    51    26    46
#>  2 Delan~ TEN     2018    57     0     0     0     0     0     0    NA     0
#>  3 Greg ~ CAR     2018    24     0     0    NA     0    98   100    97    97
#>  4 Jared~ NO      2018    91    80    82    83    91    85     0    62    60
#>  5 Marce~ GB      2018    12    18    10    22    15    31    NA    23     9
#>  6 Verno~ WAS     2018    59    36    61    NA    43    51    45    49    46
#>  7 Ed Di~ SEA     2018     0     0     0     0     0     0    NA    32    31
#>  8 Jimmy~ GB      2018    98    75    93    71    81    79    NA    88    80
#>  9 Logan~ ATL     2018    27    46    31    41    26    55    42    NA    44
#> 10 Kyle ~ MIN     2018    83    86    97    81    87    85    81    86    94
#> # ... with 100 more rows, and 10 more variables: w10 <dbl>, w11 <dbl>,
#> #   w12 <dbl>, w13 <dbl>, w14 <dbl>, w15 <dbl>, w16 <dbl>, w17 <dbl>,
#> #   ttl <dbl>, avg <dbl>

# all defensive positions for the current NFL season.
fp_snap_counts(pos = "defense")

#> # A tibble: 769 x 23
#>    player pos   team  season    w1    w2    w3    w4    w5    w6    w7    w8
#>    <chr>  <chr> <chr>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Patri~ CB    ARI     2019    78    72    73    65    95    71    58    66
#>  2 Adria~ DE    ATL     2019    22    21    31    18    42    18    33    20
#>  3 Princ~ CB    CHI     2019    59    66    25     0    NA    52    63    54
#>  4 Marce~ DT    JAC     2019    46    46    42    32    46    44    34    31
#>  5 Camer~ DE    PIT     2019    76    50    54    56    51    49    NA    53
#>  6 Von M~ LB    DEN     2019    50    50    49    53    46    52    56    52
#>  7 Rober~ DE    DAL     2019    41    33    54    45    41    50    37    31
#>  8 J.J. ~ DE    HOU     2019    70    59    52    79    56    54    61    53
#>  9 Justi~ DE    IND     2019    63    70    64    58    22     0     0     0
#> 10 Ryan ~ LB    WAS     2019    36    50    57    NA    48    48    49    58
#> # ... with 759 more rows, and 11 more variables: w9 <dbl>, w10 <dbl>,
#> #   w11 <dbl>, w12 <dbl>, w13 <dbl>, w14 <dbl>, w15 <dbl>, w16 <dbl>,
#> #   w17 <dbl>, ttl <dbl>, avg <dbl>