NFL Positional Statistics

NFL Positional Statistics

fp_stats(pos = "QB", season, start_week = 1, end_week = 17,
  scoring = c("half", "ppr", "std"))

Arguments

pos
Charcater. Specific position you want to return. Default will return all all offensive postions. Available options include

"QB"

"RB"

"WR"

"TE"

"K"

"DST"

"DL"

"LB"

"DB"

season
Numeric. The NFL season. If missing it will return the current year's season. Supported season only go back to 2015

start_week
Numeric. The starting week. Default is 1

end_week
Numeric. The ending week. Default is 17

scoring
Charcater. Fantasy scoring format. Default is "half"

Value

a tibble

Note

If you query a single week all players that were on a bye that week are not returned

Examples

# overall fantasy points for the entire season
fp_stats("RB", season = 2018)

#> # A tibble: 332 x 23
#>    player pos   team  season start_week end_week scoring rushing_att rushing_yds
#>    <chr>  <chr> <chr>  <dbl>      <dbl>    <dbl> <chr>         <dbl>       <dbl>
#>  1 Aaron~ RB    ATL     2018          1       17 half              0           0
#>  2 Adria~ RB    WAS     2018          1       17 half            251        1042
#>  3 Ahman~ RB    GB      2018          1       17 half              0           0
#>  4 Alber~ RB    JAC     2018          1       17 half              0           0
#>  5 Alvin~ RB    TEN     2018          1       17 half              0           0
#>  6 Antho~ RB    WAS     2018          1       17 half              0           0
#>  7 Anton~ RB    LAR     2018          1       17 half              0           0
#>  8 B.J. ~ RB    TB      2018          1       17 half              0           0
#>  9 Brad ~ RB    CAR     2018          1       17 half              0           0
#> 10 Carey~ RB    WAS     2018          1       17 half              0           0
#> # ... with 322 more rows, and 14 more variables: rushing_y_a <dbl>,
#> #   rushing_lg <dbl>, rushing_20 <dbl>, rushing_td <dbl>, receiving_rec <dbl>,
#> #   receiving_tgt <dbl>, receiving_yds <dbl>, receiving_y_r <dbl>,
#> #   receiving_td <dbl>, fl <dbl>, g <dbl>, fpts <dbl>, fpts_g <dbl>, own <dbl>

# TE's in the 2017 season using standard scoring
fp_stats(pos = "TE", season = 2017, scoring = "std")

#> # A tibble: 285 x 22
#>    player pos   team  season start_week end_week scoring receiving_rec
#>    <chr>  <chr> <chr>  <dbl>      <dbl>    <dbl> <chr>           <dbl>
#>  1 Alge ~ TE    NE      2017          1       17 std                 0
#>  2 Antho~ TE    FA      2017          1       17 std                12
#>  3 Anton~ TE    FA      2017          1       17 std                30
#>  4 Ben P~ TE    NYG     2017          1       17 std                 0
#>  5 Benja~ TE    NE      2017          1       17 std                61
#>  6 Billy~ TE    NO      2017          1       17 std                 0
#>  7 Brad ~ TE    KC      2017          1       17 std                 0
#>  8 Brad ~ TE    CIN     2017          1       17 std                 0
#>  9 Brent~ TE    FA      2017          1       17 std                13
#> 10 Buck ~ TE    NO      2017          1       17 std                 0
#> # ... with 275 more rows, and 14 more variables: receiving_tgt <dbl>,
#> #   receiving_yds <dbl>, receiving_y_r <dbl>, receiving_lg <dbl>,
#> #   receiving_20 <dbl>, receiving_td <dbl>, rushing_att <dbl>,
#> #   rushing_yds <dbl>, rushing_td <dbl>, fl <dbl>, g <dbl>, fpts <dbl>,
#> #   fpts_g <dbl>, own <dbl>